--- a/package.json
+++ b/package.json
@@ -9,7 +9,8 @@
     "test:lint:js": "eslint test/ --ext .js",
     "test:lint:html": "eslint *.html",
     "test:lint": "npm run test:lint:js && npm run test:lint:html",
-    "test": "npm run test:lint && wct --skip-plugin sauce"
+    "test:no-lint": "wct --skip-plugin sauce -p",
+    "test": "npm run test:lint && npm run test:no-lint"
   },
   "author": "D2L Corporation",
   "license": "Apache-2.0",