@ryantmer 2017-01-11
For posterity, I `bower link`ed this to My Courses, which uses a lot of these elements, and all seems to work fine. No real noticeable perf difference unfortunately, but there is lots of other heavy stuff there that is likely hiding any gains we would make here anyway.