@ryantmer 2017-01-11
Hm, although does cause some My Courses tests to fail... I'll check what's up with that, although might require a major version bump here (which like... I guess _technically_ it is, in a very specific way).