@ryantmer 2017-01-11
Turns out a default value is required for the observed values... or something. If they don't all have a default value, the observer isn't called. Fortunately adding default values to all three doesn't seem to have much of a performance impact, and we're still at ~170-180ms/250 in polyperf.