@apalaniuk 2017-01-11
@ryantmer Yeah, multi-property observers aren't called until all dependent properties are defined, so any method that ensures that will work/is necessary.