@ryantmer 2017-01-11
@apalaniuk doh! I had no idea that was the case, [despite it being right in the docs](https://www.polymer-project.org/1.0/docs/devguide/observers#multi-property-observers). Thanks!